more debugging options

+++ app/Http/Controllers/UsersController.php
@@ -193,40 +193,44 @@ class UsersController extends Controller
         Log::debug("first check: " . print_r($data, true));
         if (isset($data->entry[0]->changes)) {
             if (isset($data->entry[0]->changes[0]->value)) {
-                $inboundNotification = isset($data->entry[0]->changes[0]->value->messages[0]);
-                $messageType = $data->entry[0]->changes[0]->value->messages[0]->type;
-                if ($inboundNotification) {
-                    $to = $data->entry[0]->changes[0]->value->messages[0]->from;
-                    $subscriber = Subscriber::where('telephone', '=', $to)->first();
-                    if ($subscriber) {
-                        if ($messageType == "text" || $messageType = "button") {
-                            if ($data->entry[0]->changes[0]->value->messages[0]->text->body === 'أوقف الخدمة') {
-                                $myservice = new MyService;
-                                $response = $myservice->sendWhatsApp($subscriber, [], 'stop_service');
-                                if (isset($response->error->message)) {
-                                    Log::debug($response->error->message);
+                if (isset($data->entry[0]->changes[0]->value->messages)) {
+                    $inboundNotification = isset($data->entry[0]->changes[0]->value->messages[0]);
+                    $messageType = $data->entry[0]->changes[0]->value->messages[0]->type;
+                    if ($inboundNotification) {
+                        $to = $data->entry[0]->changes[0]->value->messages[0]->from;
+                        $subscriber = Subscriber::where('telephone', '=', $to)->first();
+                        if ($subscriber) {
+                            if ($messageType == "text" || $messageType = "button") {
+                                if ($data->entry[0]->changes[0]->value->messages[0]->text->body === 'أوقف الخدمة') {
+                                    $myservice = new MyService;
+                                    $response = $myservice->sendWhatsApp($subscriber, [], 'stop_service');
+                                    if (isset($response->error->message)) {
+                                        Log::debug($response->error->message);
+                                    } else {
+                                        Log::debug("One message was sent to one user who sent request to stop service");
+                                    }
+                                    $subscriber = Subscriber::where('telephone', '=', $to);
+                                    $subscriber->delete();
                                 } else {
-                                    Log::debug("One message was sent to one user who sent request to stop service");
+                                    $myservice = new MyService;
+                                    $response = $myservice->sendWhatsApp($subscriber, [], 'to_stop_service');
+                                    if (isset($response->error->message)) {
+                                        Log::debug($response->error->message);
+                                    } else {
+                                        Log::debug("One message was sent to one user who sent something");
+                                    }
                                 }
-                                $subscriber = Subscriber::where('telephone', '=', $to);
-                                $subscriber->delete();
                             } else {
-                                $myservice = new MyService;
-                                $response = $myservice->sendWhatsApp($subscriber, [], 'to_stop_service');
-                                if (isset($response->error->message)) {
-                                    Log::debug($response->error->message);
-                                } else {
-                                    Log::debug("One message was sent to one user who sent something");
-                                }
+                                // message is not text
                             }
                         } else {
-                            // message is not text
+                            // user no longer a subscriber
                         }
                     } else {
-                        // user no longer a subscriber
+                        Log::debug(print_r($data, true));
                     }
                 } else {
-                    Log::debug(print_r($data, true));
+                    Log::debug("message status is: " . print_r($data->entry[0]->changes[0]->value->statuses[0]->status, true));
                 }
             } else {
                 Log::debug(print_r($data, true));
